+++ app/Http/Controllers/AuthController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\LoginUserRequest;
+use App\Http\Requests\RegisterUserRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -76,5 +77,20 @@ class AuthController extends Controller
             $user->email = $userGoogle->email;
         }*/
 
+    }
+    public function showPageRegister()
+    {
+        return view('register');
+    }
+
+    public function register(RegisterUserRequest $request)
+    {
+        $user = new User();
+        $user->username = $request->username;
+        $user->phone = $request->phone;
+        $user->password = hash("md5", $request->password);
+        $user->email = $request->email;
+        $user->save();
+        return redirect()->route('page.login');
     }
 }
